Bouge le statut de la demande de sub dans le domaine
