+++ src/gateways/PrismaLesFeuillesDeRouteLoader.ts
@@ -5,7 +5,8 @@ import { Membre, membreInclude, toMembre, toMembres } from './shared/MembresGouv
 import prisma from '../../prisma/prismaClient'
 import { alphaAsc } from '@/shared/lang'
 import { FeuillesDeRouteLoader, FeuillesDeRouteReadModel } from '@/use-cases/queries/RecupererLesFeuillesDeRoute'
-import { BesoinsPossible, StatutSubvention } from '@/use-cases/queries/shared/ActionReadModel'
+import { BesoinsPossible } from '@/use-cases/queries/shared/ActionReadModel'
+import { StatutSubvention } from "@/domain/DemandeDeSubvention"
 import { EtablisseurSyntheseGouvernance } from '@/use-cases/services/shared/etablisseur-synthese-gouvernance'
 
 export class PrismaLesFeuillesDeRouteLoader implements FeuillesDeRouteLoader {
