+++ src/use-cases/commands/AjouterUneAction.ts
@@ -8,10 +8,9 @@ import { GetFeuilleDeRouteRepository, UpdateFeuilleDeRouteRepository } from './s
 import { GetGouvernanceRepository } from './shared/GouvernanceRepository'
 import { TransactionRepository } from './shared/TransactionRepository'
 import { GetUtilisateurRepository } from './shared/UtilisateurRepository'
-import { StatutSubvention } from '../queries/shared/ActionReadModel'
 import { Action, ActionFailure } from '@/domain/Action'
 import { CoFinancement, CoFinancementFailure } from '@/domain/CoFinancement'
-import { DemandeDeSubvention, DemandeDeSubventionFailure } from '@/domain/DemandeDeSubvention'
+import { DemandeDeSubvention , DemandeDeSubventionFailure, StatutSubvention } from '@/domain/DemandeDeSubvention'
 import { FeuilleDeRoute } from '@/domain/FeuilleDeRoute'
 import { GouvernanceUid } from '@/domain/Gouvernance'
 import { Utilisateur } from '@/domain/Utilisateur'
