+++ src/use-cases/queries/RecupererLesFeuillesDeRoute.ts
@@ -1,4 +1,5 @@
-import { BesoinsPossible, StatutSubvention } from './shared/ActionReadModel'
+import { BesoinsPossible } from './shared/ActionReadModel'
+import { StatutSubvention } from "@/domain/DemandeDeSubvention"
 
 export interface FeuillesDeRouteLoader {
   get(codeDepartement: string): Promise<FeuillesDeRouteReadModel>
