+++ src/use-cases/services/EtablirSyntheseFinanciereGouvernance.ts
@@ -1,5 +1,5 @@
 import { Finances, Gouvernance, SyntheseGouvernance } from './shared/etablisseur-synthese-gouvernance'
-import { StatutSubvention } from '../queries/shared/ActionReadModel'
+import { StatutSubvention } from "@/domain/DemandeDeSubvention"
 
 export function etablirSyntheseFinanciereGouvernance(gouvernance: Gouvernance): SyntheseGouvernance {
   const bilanGouvernance = gouvernance.feuillesDeRoute.reduce(
